ci: measure explicit conversion preview diagnostics

+++ engine/botlib/be_aas_move.cpp
@@ -316,7 +316,7 @@ static float AAS_WeaponJumpZVelocity(vec3_t origin, float radiusdamage)
 	VectorMA(origin, 0.5, v, v);
 	VectorSubtract(bsptrace.endpos, v, v);
 	//
-	points = radiusdamage - 0.5 * VectorLength(v);
+	points = (float)( radiusdamage - 0.5 * VectorLength(v) );
 	if (points < 0) points = 0;
 	//the owner of the rocket gets half the damage
 	points *= 0.5;
@@ -394,7 +394,7 @@ static void AAS_ApplyFriction(vec3_t vel, float friction, float stopspeed,
 	float speed, control, newspeed;
 
 	//horizontal speed
-	speed = sqrt((double)(vel[0] * vel[0] + vel[1] * vel[1]));
+	speed = (float)( sqrt((double)(vel[0] * vel[0] + vel[1] * vel[1])) );
 	if (speed)
 	{
 		control = speed < stopspeed ? stopspeed : speed;
@@ -551,7 +551,7 @@ static int AAS_ClientMovementPrediction( aas_clientmove_t *move,
 		//get gravity depending on swimming or not
 		gravity = swimming ? phys_watergravity : phys_gravity;
 		//apply gravity at the START of the frame
-		frame_test_vel[2] = frame_test_vel[2] - (gravity * 0.1 * frametime);
+		frame_test_vel[2] = (float)( frame_test_vel[2] - (gravity * 0.1 * frametime) );
 		//if on the ground or swimming
 		if (onground || swimming)
 		{
@@ -580,7 +580,7 @@ static int AAS_ClientMovementPrediction( aas_clientmove_t *move,
 				if (!swimming && cmdmove[2] > 1)
 				{
 					//jump velocity minus the gravity for one frame + 5 for safety
-					frame_test_vel[2] = phys_jumpvel - (gravity * 0.1 * frametime) + 5;
+					frame_test_vel[2] = (float)( phys_jumpvel - (gravity * 0.1 * frametime) + 5 );
 					jump_frame = n;
 					//jumping so air accelerate
 					accelerate = phys_airaccelerate;
@@ -838,7 +838,7 @@ static int AAS_ClientMovementPrediction( aas_clientmove_t *move,
 						if (delta)
 						{
 							delta = delta * 10;
-							delta = delta * delta * 0.0001;
+							delta = (float)( delta * delta * 0.0001 );
 							if (swimming) delta = 0;
 							// never take falling damage if completely underwater
 							/*
@@ -1063,7 +1063,7 @@ int AAS_HorizontalVelocityForJump(float zvel, vec3_t start, vec3_t end, float *v
 	phys_maxvelocity = aassettings.phys_maxvelocity;
 
 	//maximum height a player can jump with the given initial z velocity
-	maxjump = 0.5 * phys_gravity * (zvel / phys_gravity) * (zvel / phys_gravity);
+	maxjump = (float)( 0.5 * phys_gravity * (zvel / phys_gravity) * (zvel / phys_gravity) );
 	//top of the parabolic jump
 	top = start[2] + maxjump;
 	//height the bot will fall from the top
@@ -1075,7 +1075,7 @@ int AAS_HorizontalVelocityForJump(float zvel, vec3_t start, vec3_t end, float *v
 		return 0;
 	} //end if
 	//time a player takes to fall the height
-	t = sqrt(height2fall / (0.5 * phys_gravity));
+	t = (float)( sqrt(height2fall / (0.5 * phys_gravity)) );
   	//direction from start to end
 	VectorSubtract(end, start, dir);
 	//
@@ -1084,7 +1084,7 @@ int AAS_HorizontalVelocityForJump(float zvel, vec3_t start, vec3_t end, float *v
 		return 0;
 	}
 	//calculate horizontal speed
-	*velocity = sqrt((double)(dir[0]*dir[0] + dir[1]*dir[1])) / (t + zvel / phys_gravity);
+	*velocity = (float)( sqrt((double)(dir[0]*dir[0] + dir[1]*dir[1])) / (t + zvel / phys_gravity) );
 	//the horizontal speed must be lower than the max speed
 	if (*velocity > phys_maxvelocity)
 	{
