+++ engine/client/cl_cgame.cpp
@@ -1106,7 +1106,7 @@ static float CL_AvgPing( void ) {
 	if ( (count % 2) == 0 )
 		result = (ping[count / 2] + ping[(count / 2) - 1]) / 2.0f;
 	else
-		result = ping[count / 2];
+		result = (float)( ping[count / 2] );
 
 	return result;
 }
